requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
langchain==0.1.224
langchain-google-genai==0.0.22
langchain-community==0.0.202
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
